결제 페이지 퍼블리싱

+++ domain/src/main/java/com/danbam/domain/entity/FundingEntity.kt
@@ -4,7 +4,7 @@ data class FundingEntity(
     val idx: Long,
     val title: String,
     val description: String,
-    val percentage: Long,
+    val percentage: Double,
     val thumbnail: String,
     val status: String,
 )
